requirements.txt
# =============================================================
# Stock Market Analysis Agent - Python Dependencies
# =============================================================
# Install with:  pip install -r requirements.txt
# =============================================================

openai>=1.12.0
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.24.0
python-dotenv>=1.0.0
schedule>=1.2.1
python-telegram-bot>=21.0
requests>=2.31.0
jinja2>=3.1.3
aiohttp>=3.9.0
fpdf2>=2.7.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
